Fractional-share sizing and $500-account levers
Adds risk.fractional_shares / fractional_decimals / min_order_notional_usd.
Equities are sized to 4 decimals (options stay whole contracts); orders
below Robinhood's $1 minimum size to zero. Retunes config/levers.yaml for a
$500 account and disables options there.

+++ tradeagent/models.py
@@ -96,7 +96,7 @@ class Proposal(BaseModel):
     confidence: float = Field(ge=0, le=1)
     tags: list[str] = Field(default_factory=list)
     evidence: list[str] = Field(default_factory=list)
-    requested_qty: Optional[int] = None
+    requested_qty: Optional[float] = None
     notes: Optional[str] = None
 
     @field_validator("symbol")
